Rewrite decode_structure; it's now called ensure_unicode.
It now does a much more thorough job of cleaning up feedparser's return
value: it processes lists properly, it uses isinstance rather than an
explicit list of dictionary types, and it tries to reduce everything to
Python's built-in types.

This was prompted by Joseph Reagle reporting that rawdog doesn't work
with Ubuntu's version of BeautifulSoup -- which returns (and leaks into
feedparser) instances of NavigableString, a subclass of unicode that
can't be pickled. These are now squashed into real unicode objects.