Completely redesign the persistance mechanism.
There's now a Persisted class that acts as a context manager for a
persisted file (and supports __enter__/__exit__ for "with"), and a
Persister class that knows about the collection of files currently open.

This was initially to fix the feed-redirection bug that resulted in
opening the same statefile twice -- but it also cleans up the main code
quite a bit.